Fix CoverEntityFeature deprecation warning
Original warning:
```
Logger: homeassistant.components.cover
Quelle: helpers/deprecation.py:222
Integration: Abdeckung (Dokumentation, Probleme)
Erstmals aufgetreten: 19:01:15 (12 Vorkommnisse)
Zuletzt protokolliert: 19:01:15

SUPPORT_CLOSE_TILT was used from hass-becker-component, this is a deprecated constant which will be removed in HA Core 2025.1. Use CoverEntityFeature.CLOSE_TILT instead, please report it to the author of the 'hass-becker-component' custom integration
SUPPORT_OPEN was used from hass-becker-component, this is a deprecated constant which will be removed in HA Core 2025.1. Use CoverEntityFeature.OPEN instead, please report it to the author of the 'hass-becker-component' custom integration
SUPPORT_OPEN_TILT was used from hass-becker-component, this is a deprecated constant which will be removed in HA Core 2025.1. Use CoverEntityFeature.OPEN_TILT instead, please report it to the author of the 'hass-becker-component' custom integration
SUPPORT_STOP was used from hass-becker-component, this is a deprecated constant which will be removed in HA Core 2025.1. Use CoverEntityFeature.STOP instead, please report it to the author of the 'hass-becker-component' custom integration
SUPPORT_SET_POSITION was used from hass-becker-component, this is a deprecated constant which will be removed in HA Core 2025.1. Use CoverEntityFeature.SET_POSITION instead, please report it to the author of the 'hass-becker-component' custom integration
```